# openapi.yml
openapi: 3.1.0
info:
  title: API de réservation de salles
  version: 1.0.0
  description: |
    Contrat de l'API de gestion des salles, équipements, organisateurs et réservations.
    Les dates et heures utilisent le format ISO 8601 avec fuseau ou décalage UTC.
servers:
  - url: http://localhost:8080
    description: Environnement local
...
paths:
  /api/buildings:
    post:
      tags: [Buildings]
      summary: Créer un bâtiment
      description: Crée un bâtiment avec un nom unique et un nombre d'étages positif.
      operationId: createBuilding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildingRequest'
      responses:
        '201':
          description: Bâtiment créé
          headers:
            Location:
              description: URI du bâtiment créé
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingResponse'
        '400':
          $ref: '#/components/responses/ValidationBadRequest'
        '409':
          $ref: '#/components/responses/ResourceAlreadyExists'
    get:
      tags: [Buildings]
      summary: Lister les bâtiments
      description: Retourne tous les bâtiments, triés par nom sans tenir compte de la casse.
      operationId: listBuildings
      responses:
        '200':
          description: Liste des bâtiments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuildingResponse'
  /api/buildings/{buildingId}:
    parameters:
      - $ref: '#/components/parameters/BuildingId'
    get:
      tags: [Buildings]
      summary: Consulter un bâtiment
      description: Retourne le bâtiment demandé et son nombre d'étages.
      operationId: getBuilding
      responses:
        '200':
          description: Bâtiment trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingResponse'
        '404':
          $ref: '#/components/responses/BuildingNotFound'
    put:
      tags: [Buildings]
      summary: Modifier un bâtiment
      description: |
        Remplace le nom et le nombre d'étages. La réduction du nombre d'étages est
        refusée lorsqu'une salle ou un organisateur occupe un étage qui deviendrait invalide.
      operationId: updateBuilding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBuildingRequest'
      responses:
        '200':
          description: Bâtiment mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingResponse'
        '400':
          $ref: '#/components/responses/ValidationBadRequest'
        '404':
          $ref: '#/components/responses/BuildingNotFound'
        '409':
          $ref: '#/components/responses/BuildingUpdateConflict'
...
  /api/equipment:
    post:
      tags: [Equipment]
      summary: Créer un équipement
      description: Crée un équipement dont le code est unique et respecte le format attendu.
      operationId: createEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEquipmentRequest'
      responses:
        '201':
          description: Équipement créé
          headers:
            Location:
              description: URI de l'équipement créé
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '400':
          $ref: '#/components/responses/ValidationBadRequest'
        '409':
          $ref: '#/components/responses/ResourceAlreadyExists'
    get:
      tags: [Equipment]
      summary: Lister les équipements
      description: Retourne tous les équipements disponibles, triés par code croissant.
      operationId: listEquipment
      responses:
        '200':
          description: Liste des équipements, triée par code
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentResponse'
  /api/organizers:
    post:
      tags: [Organizers]
      summary: Créer un organisateur
      description: |
        Crée un organisateur avec une adresse e-mail valide et unique, la référence d'un
        bâtiment existant et un étage valide dans ce bâtiment. Cette localisation sert
        au calcul de proximité avec les salles.
      operationId: createOrganizer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizerRequest'
      responses:
        '201':
          description: Organisateur créé
          headers:
            Location:
              description: URI de l'organisateur créé
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerResponse'
        '400':
          $ref: '#/components/responses/ValidationBadRequest'
        '404':
          $ref: '#/components/responses/OrganizerBuildingNotFound'
        '409':
          $ref: '#/components/responses/ResourceAlreadyExists'

    get:
      tags: [Organizers]
      summary: Lister les organisateurs
      description: Retourne tous les organisateurs, triés par nom puis par identifiant.
      operationId: listOrganizers
      responses:
        '200':
          description: Liste des organisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizerResponse'
  /api/organizers/{organizerId}:
    parameters:
      - $ref: '#/components/parameters/OrganizerId'
    get:
      tags: [Organizers]
      summary: Consulter un organisateur
      description: Retourne l'identité, l'adresse e-mail et la localisation de l'organisateur.
      operationId: getOrganizer
      responses:
        '200':
          description: Organisateur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerResponse'
        '404':
          $ref: '#/components/responses/OrganizerNotFound'
...
components:
  parameters:
    BuildingId:
      name: buildingId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
...
    OrganizerId:
      name: organizerId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
...
  responses:
    ValidationBadRequest:
      description: Un ou plusieurs champs de la requête sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    BuildingNotFound:
      description: Le bâtiment demandé n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: BUILDING_NOT_FOUND
            message: Le bâtiment 3 n'existe pas
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/buildings/3
            details:
              buildingId: 3
            fieldErrors: {}
    OrganizerBuildingNotFound:
      description: Le bâtiment associé à l'organisateur n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: BUILDING_NOT_FOUND
            message: Le bâtiment 3 n'existe pas
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/organizers
            details:
              buildingId: 3
            fieldErrors: {}
...
    OrganizerNotFound:
      description: L'organisateur demandé n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: ORGANIZER_NOT_FOUND
            message: L'organisateur 12 n'existe pas
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/organizers/12
            details:
              organizerId: 12
            fieldErrors: {}
...
    ResourceAlreadyExists:
      description: Une valeur soumise à une contrainte d'unicité est déjà utilisée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    BuildingUpdateConflict:
      description: Le nom est déjà utilisé ou le nouveau nombre d'étages est incompatible
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            nameAlreadyExists:
              value:
                code: RESOURCE_ALREADY_EXISTS
                message: Un bâtiment utilise déjà ce nom
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/buildings/3
                details: {}
                fieldErrors: {}
            occupiedFloor:
              value:
                code: BUILDING_FLOOR_COUNT_CONFLICT
                message: Une salle ou un organisateur occupe un étage qui serait supprimé
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/buildings/3
                details:
                  highestOccupiedFloor: 4
                fieldErrors: {}
...
  schemas:
    ResourceId:
      type: integer
      format: int64
      minimum: 1
      example: 7
...
    EquipmentCode:
      type: string
      pattern: '^[A-Z][A-Z0-9_]{1,49}$'
      example: PROJECTOR
    BuildingName:
      type: string
      minLength: 1
      maxLength: 100
      description: Nom unique du bâtiment, sans tenir compte de la casse
      example: Bâtiment A
    FloorNumber:
      type: integer
      format: int32
      minimum: 0
      maximum: 199
      description: Doit être inférieur au nombre d'étages du bâtiment référencé
      example: 2
    NumberOfFloors:
      type: integer
      format: int32
      minimum: 1
      maximum: 200
      description: Nombre d'étages, rez-de-chaussée inclus
      example: 5
    CreateBuildingRequest:
      type: object
      additionalProperties: false
      required: [name, numberOfFloors]
      properties:
        name:
          $ref: '#/components/schemas/BuildingName'
        numberOfFloors:
          $ref: '#/components/schemas/NumberOfFloors'
    UpdateBuildingRequest:
      type: object
      additionalProperties: false
      required: [name, numberOfFloors]
      properties:
        name:
          $ref: '#/components/schemas/BuildingName'
        numberOfFloors:
          $ref: '#/components/schemas/NumberOfFloors'
    BuildingResponse:
      type: object
      additionalProperties: false
      required: [id, name, numberOfFloors]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          $ref: '#/components/schemas/BuildingName'
        numberOfFloors:
          $ref: '#/components/schemas/NumberOfFloors'
...
    CreateEquipmentRequest:
      type: object
      additionalProperties: false
      required: [code, label]
      properties:
        code:
          $ref: '#/components/schemas/EquipmentCode'
        label:
          type: string
          minLength: 1
          maxLength: 100
          example: Vidéoprojecteur
    EquipmentResponse:
      type: object
      additionalProperties: false
      required: [id, code, label]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        code:
          $ref: '#/components/schemas/EquipmentCode'
        label:
          type: string
          example: Vidéoprojecteur
    CreateOrganizerRequest:
      type: object
      additionalProperties: false
      required: [name, email, buildingId, floor]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
          example: Alice Martin
        email:
          type: string
          format: email
          maxLength: 254
          example: alice.martin@example.org
        buildingId:
          $ref: '#/components/schemas/ResourceId'
        floor:
          $ref: '#/components/schemas/FloorNumber'
    OrganizerSummaryResponse:
      type: object
      required: [id, name, building, floor]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          example: Alice Martin
        building:
          $ref: '#/components/schemas/BuildingResponse'
        floor:
          $ref: '#/components/schemas/FloorNumber'
    OrganizerResponse:
      allOf:
        - $ref: '#/components/schemas/OrganizerSummaryResponse'
        - type: object
          required: [email]
          properties:
            email:
              type: string
              format: email
              example: alice.martin@example.org
...
    ApiErrorResponse:
      type: object
      description: |
        Format commun à toutes les erreurs de l'API.

        | Code | HTTP | Signification |
        | --- | ---: | --- |
        | `VALIDATION_ERROR` | 400 | Un ou plusieurs champs sont invalides. |
        | `INVALID_RESERVATION_PERIOD` | 400 | La période est passée, inversée ou dépasse huit heures. |
        | `BUILDING_NOT_FOUND` | 404 | Le bâtiment demandé n'existe pas. |
        | `ROOM_NOT_FOUND` | 404 | La salle demandée n'existe pas. |
        | `ORGANIZER_NOT_FOUND` | 404 | L'organisateur demandé n'existe pas. |
        | `RESERVATION_NOT_FOUND` | 404 | La réservation demandée n'existe pas. |
        | `EQUIPMENT_NOT_FOUND` | 404 | Au moins un code d'équipement n'existe pas. |
        | `ROOM_CAPACITY_EXCEEDED` | 409 | La salle choisie est trop petite. |
        | `MISSING_REQUIRED_EQUIPMENT` | 409 | La salle choisie ne possède pas tous les équipements demandés. |
        | `ROOM_ALREADY_RESERVED` | 409 | Une réservation confirmée chevauche la période. |
        | `ROOM_UNAVAILABLE` | 409 | La salle choisie est en maintenance. |
        | `NO_COMPATIBLE_ROOM` | 409 | Aucune salle ne convient à l'attribution automatique. |
        | `RESERVATION_ALREADY_CANCELLED` | 409 | La réservation est déjà annulée. |
        | `RESOURCE_ALREADY_EXISTS` | 409 | Un nom, code ou e-mail unique est déjà utilisé. |
        | `BUILDING_FLOOR_COUNT_CONFLICT` | 409 | Le nombre d'étages est incompatible avec une localisation existante. |
      additionalProperties: false
      required: [code, message, timestamp, path]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - INVALID_RESERVATION_PERIOD
            - BUILDING_NOT_FOUND
            - ROOM_NOT_FOUND
            - ORGANIZER_NOT_FOUND
            - RESERVATION_NOT_FOUND
            - EQUIPMENT_NOT_FOUND
            - ROOM_CAPACITY_EXCEEDED
            - MISSING_REQUIRED_EQUIPMENT
            - ROOM_ALREADY_RESERVED
            - ROOM_UNAVAILABLE
            - NO_COMPATIBLE_ROOM
            - RESERVATION_ALREADY_CANCELLED
            - RESOURCE_ALREADY_EXISTS
            - BUILDING_FLOOR_COUNT_CONFLICT
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        details:
          type: object
          additionalProperties: true
          default: {}
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          default: {}